+++ src/model-cache/tests/ModelCacheTest.php
@@ -43,9 +43,11 @@ class ModelCacheTest extends TestCase
     {
         ContainerStub::mockContainer();
 
-        $users = UserModel::findManyFromCache([1, 2]);
-        $expects = UserModel::query()->findMany([1, 2]);
+        $users = UserModel::findManyFromCache([1, 2, 3]);
+        $expects = UserModel::query()->findMany([1, 2, 3]);
 
+        $this->assertTrue(count($users) == 2);
+        $this->assertEquals([1, 2], array_keys($users->getDictionary()));
         $this->assertEquals($expects, $users);
     }
 
@@ -107,5 +109,68 @@ class ModelCacheTest extends TestCase
         }
 
         UserModel::query(true)->whereIn('id', $ids)->delete();
+    }
+
+    public function testIncr()
+    {
+        $container = ContainerStub::mockContainer();
+
+        $id = 206;
+        UserModel::query()->firstOrCreate(['id' => $id], [
+            'name' => uniqid(),
+            'gender' => 1,
+        ]);
+
+        $model = UserModel::findFromCache($id);
+        /** @var \Redis $redis */
+        $redis = $container->make(RedisProxy::class, ['pool' => 'default']);
+        $this->assertEquals(1, $redis->exists('{mc:default:m:user}:id:' . $id));
+
+        $this->assertEquals(1, $model->increment('gender', 1));
+        $this->assertEquals(1, $redis->exists('{mc:default:m:user}:id:' . $id));
+        $this->assertEquals(2, $redis->hGet('{mc:default:m:user}:id:' . $id, 'gender'));
+        $this->assertEquals(2, UserModel::findFromCache($id)->gender);
+        $this->assertEquals(2, UserModel::query()->find($id)->gender);
+
+        UserModel::query(true)->where('id', $id)->delete();
+    }
+
+    public function testFindNullBeforeCreate()
+    {
+        $container = ContainerStub::mockContainer();
+
+        $id = 207;
+
+        $model = UserModel::findFromCache($id);
+        /** @var \Redis $redis */
+        $redis = $container->make(RedisProxy::class, ['pool' => 'default']);
+        $this->assertEquals(1, $redis->exists('{mc:default:m:user}:id:' . $id));
+        $this->assertNull($model);
+
+        $this->assertEquals(1, $redis->del('{mc:default:m:user}:id:' . $id));
+        UserModel::query(true)->where('id', $id)->delete();
+    }
+
+    public function testIncrNotExist()
+    {
+        $container = ContainerStub::mockContainer();
+
+        $id = 206;
+        UserModel::query()->firstOrCreate(['id' => $id], [
+            'name' => uniqid(),
+            'gender' => 1,
+        ]);
+
+        $model = UserModel::query()->find($id);
+        /** @var \Redis $redis */
+        $redis = $container->make(RedisProxy::class, ['pool' => 'default']);
+        $this->assertEquals(0, $redis->exists('{mc:default:m:user}:id:' . $id));
+
+        $this->assertEquals(1, $model->increment('gender', 1));
+        $this->assertEquals(0, $redis->exists('{mc:default:m:user}:id:' . $id));
+        $this->assertEquals(2, UserModel::query()->find($id)->gender);
+        $this->assertEquals(2, UserModel::findFromCache($id)->gender);
+
+        UserModel::query(true)->where('id', $id)->delete();
     }
 }